feat: Nested dialogs via history module
