+++ panel/src/panel/modal.js
@@ -2,9 +2,11 @@
 
 import { isObject } from "@/helpers/object.js";
 import Feature, { defaults as featureDefaults } from "./feature.js";
+import History from "./history.js";
 import focus from "@/helpers/focus.js";
 import "@/helpers/array.js";
 import { reactive, set } from "vue";
+import { uuid } from "@/helpers/string.js";
 import { wrap } from "@/helpers/array.js";
 
 /**
@@ -12,7 +14,8 @@ import { wrap } from "@/helpers/array.js";
  */
 export const defaults = () => {
 	return {
-		...featureDefaults()
+		...featureDefaults(),
+		id: null
 	};
 };
 
@@ -47,22 +50,43 @@ export default (panel, key, defaults) => {
 		},
 
 		/**
-		 * Closes the modal
+		 * Closes the modal and goes back to the
+		 * parent one if it has been stored
+		 * @param {String|true} id Which modal to close, true for all
 		 */
-		async close() {
+		async close(id) {
 			if (this.isOpen === false) {
 				return;
 			}
 
-			this.isOpen = false;
-			this.emit("close");
-			this.reset();
-
-			if (panel.overlays().length === 0) {
-				// unblock the overflow until we can use :has for this.
-				document.documentElement.removeAttribute("data-overlay");
-				document.documentElement.style.removeProperty("--scroll-top");
+			// Compare the modal id to avoid closing
+			// the wrong modal. This is particularly useful
+			// in nested modals.
+			if (id !== undefined && id !== true && id !== this.id) {
+				return;
 			}
+
+			if (id === true) {
+				this.history.clear();
+			} else {
+				this.history.removeLast();
+			}
+
+			// no more items in the history
+			if (this.history.isEmpty() === true) {
+				this.isOpen = false;
+				this.emit("close");
+				this.reset();
+
+				if (panel.overlays().length === 0) {
+					// unblock the overflow until we can use :has for this.
+					document.documentElement.removeAttribute("data-overlay");
+					document.documentElement.style.removeProperty("--scroll-top");
+				}
+				return;
+			}
+
+			return this.open(this.history.last());
 		},
 
 		/**
@@ -75,6 +99,16 @@ export default (panel, key, defaults) => {
 		focus(input) {
 			focus(`.k-${this.key()}-portal`, input);
 		},
+
+		goTo(id) {
+			const state = this.history.goto(id);
+
+			if (state !== undefined) {
+				this.open(state);
+			}
+		},
+
+		history: History(),
 
 		/**
 		 * Form drawers and dialogs can use this
@@ -142,6 +176,22 @@ export default (panel, key, defaults) => {
 				// mark the modal as open
 				this.isOpen = true;
 			}
+
+			return this.state();
+		},
+
+		/**
+		 * Sets a new active state for the modal
+		 * This is done whenever the state is an object
+		 * and not undefined or null
+		 *
+		 * @param {Object} state
+		 */
+		set(state) {
+			parent.set.call(this, state);
+
+			// create a unique ID for the drawer if it does not have one
+			this.id ??= uuid();
 
 			return this.state();
 		},
